Move offline backend tests to unit/
