+++ tests/integration/backend_test.py
@@ -19,7 +19,7 @@ from typing import cast, Callable, Any  # pylint: disable=unused-import
 import json
 import os
 import time
-from hypothesis import given, settings, strategies
+from hypothesis import given, settings
 import numpy as np
 import pytest
 import hypothesis.strategies as st
@@ -58,40 +58,16 @@ from pytket.extensions.quantinuum.backends.quantinuum import (
 from pytket.extensions.quantinuum.backends.api_wrappers import (
     QuantinuumAPIError,
     QuantinuumAPI,
-    QuantinuumAPIOffline,
 )
 from pytket.backends.status import StatusEnum
 from pytket.wasm import WasmFileHandler
+
 
 skip_remote_tests: bool = os.getenv("PYTKET_RUN_REMOTE_TESTS") is None
 
 REASON = (
     "PYTKET_RUN_REMOTE_TESTS not set (requires configuration of Quantinuum username)"
 )
-
-ALL_QUANTUM_HARDWARE_NAMES = [
-    "H1-1",
-    "H1-2",
-    "H2-1",
-]
-
-ALL_SIMULATOR_NAMES = [
-    "H1-1E",
-    "H1-2E",
-    "H2-1E",
-]
-
-ALL_SYNTAX_CHECKER_NAMES = [
-    "H1-1SC",
-    "H1-2SC",
-    "H2-1SC",
-]
-
-ALL_DEVICE_NAMES = [
-    *ALL_QUANTUM_HARDWARE_NAMES,
-    *ALL_SIMULATOR_NAMES,
-    *ALL_SYNTAX_CHECKER_NAMES,
-]
 
 
 @pytest.mark.parametrize("authenticated_quum_backend", [None], indirect=True)
@@ -132,43 +108,6 @@ def test_quantinuum(
         assert backend.backend_info is None
 
 
-def test_quantinuum_offline() -> None:
-    qapioffline = QuantinuumAPIOffline()
-    backend = QuantinuumBackend(
-        device_name="H1-1", machine_debug=False, api_handler=qapioffline  # type: ignore
-    )
-    c = Circuit(4, 4, "test 1")
-    c.H(0)
-    c.CX(0, 1)
-    c.Rz(0.3, 2)
-    c.CSWAP(0, 1, 2)
-    c.CRz(0.4, 2, 3)
-    c.CY(1, 3)
-    c.ZZPhase(0.1, 2, 0)
-    c.Tdg(3)
-    c.measure_all()
-    c = backend.get_compiled_circuit(c)
-    n_shots = 4
-    _ = backend.process_circuits([c], n_shots)[0]
-    expected_result = {
-        "name": "test 1",
-        "count": 4,
-        "machine": "H1-1",
-        "language": "OPENQASM 2.0",
-        "program": "...",  # not checked
-        "priority": "normal",
-        "options": {"simulator": "state-vector", "error-model": True, "tket": {}},
-    }
-    result = qapioffline.get_jobs()
-    assert result is not None
-    assert result[0]["name"] == expected_result["name"]
-    assert result[0]["count"] == expected_result["count"]
-    assert result[0]["machine"] == expected_result["machine"]
-    assert result[0]["language"] == expected_result["language"]
-    assert result[0]["priority"] == expected_result["priority"]
-    # assert result[0]["options"] == expected_result["options"]
-
-
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 @pytest.mark.parametrize(
     "authenticated_quum_backend", [{"device_name": "H1-1SC"}], indirect=True
@@ -194,50 +133,6 @@ def test_max_classical_register(
 
     with pytest.raises(CircuitNotValidError):
         backend._check_all_circuits([c])
-
-
-def test_max_classical_register_ii() -> None:
-    qapioffline = QuantinuumAPIOffline()
-    backend = QuantinuumBackend(
-        device_name="H1-1", machine_debug=False, api_handler=qapioffline  # type: ignore
-    )
-
-    c = Circuit(4, 4, "test 1")
-    c.H(0)
-    c.CX(0, 1)
-    c.measure_all()
-    c = backend.get_compiled_circuit(c)
-    assert backend._check_all_circuits([c])
-    for i in range(0, 20):
-        c.add_c_register(f"creg-{i}", 32)
-
-    assert backend._check_all_circuits([c])
-
-    for i in range(20, 200):
-        c.add_c_register(f"creg-{i}", 32)
-
-    with pytest.raises(CircuitNotValidError):
-        backend._check_all_circuits([c])
-
-
-def test_tket_pass_submission() -> None:
-    backend = QuantinuumBackend(device_name="H1-1SC", machine_debug=True)
-
-    sequence_pass = SequencePass(
-        [
-            OptimisePhaseGadgets(),
-            FullPeepholeOptimise(),
-            FullPeepholeOptimise(allow_swaps=False),
-            RemoveRedundancies(),
-        ]
-    )
-
-    c = Circuit(4, 4, "test 1")
-    c.H(0)
-    c.measure_all()
-    c = backend.get_compiled_circuit(c)
-    n_shots = 4
-    backend.process_circuits([c], n_shots, pytketpass=sequence_pass)
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
@@ -291,7 +186,7 @@ def test_multireg(
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 @pytest.mark.parametrize(
     "authenticated_quum_backend",
-    [{"device_name": name} for name in ALL_SYNTAX_CHECKER_NAMES],
+    [{"device_name": name} for name in pytest.ALL_SYNTAX_CHECKER_NAMES],
     indirect=True,
 )
 def test_default_pass(
@@ -332,7 +227,7 @@ def test_default_pass(
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 @pytest.mark.parametrize(
     "authenticated_quum_backend",
-    [{"device_name": name, "label": "test cancel"} for name in ALL_SIMULATOR_NAMES],
+    [{"device_name": name, "label": "test cancel"} for name in pytest.ALL_SIMULATOR_NAMES],
     indirect=True,
 )
 def test_cancel(
@@ -392,7 +287,7 @@ def circuits(
     "authenticated_quum_backend",
     [
         {"device_name": name}
-        for name in [*ALL_QUANTUM_HARDWARE_NAMES, *ALL_SYNTAX_CHECKER_NAMES]
+        for name in [*pytest.ALL_QUANTUM_HARDWARE_NAMES, *pytest.ALL_SYNTAX_CHECKER_NAMES]
     ],
     indirect=True,
 )
@@ -432,7 +327,7 @@ def test_cost_estimate(
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 @pytest.mark.parametrize(
     "authenticated_quum_backend",
-    [{"device_name": name} for name in ALL_SYNTAX_CHECKER_NAMES],
+    [{"device_name": name} for name in pytest.ALL_SYNTAX_CHECKER_NAMES],
     indirect=True,
 )
 def test_classical(
@@ -480,7 +375,7 @@ def test_classical(
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 @pytest.mark.parametrize(
     "authenticated_quum_backend",
-    [{"device_name": name} for name in ALL_SYNTAX_CHECKER_NAMES],
+    [{"device_name": name} for name in pytest.ALL_SYNTAX_CHECKER_NAMES],
     indirect=True,
 )
 def test_postprocess(
@@ -504,34 +399,6 @@ def test_postprocess(
     assert len(shots) == 10
 
 
-@given(
-    n_shots=strategies.integers(min_value=1, max_value=10),  # type: ignore
-    n_bits=strategies.integers(min_value=0, max_value=10),
-)
-def test_shots_bits_edgecases(n_shots, n_bits) -> None:
-    quantinuum_backend = QuantinuumBackend("H1-1SC", machine_debug=True)
-    c = Circuit(n_bits, n_bits)
-
-    # TODO TKET-813 add more shot based backends and move to integration tests
-    h = quantinuum_backend.process_circuit(c, n_shots)
-    res = quantinuum_backend.get_result(h)
-
-    correct_shots = np.zeros((n_shots, n_bits), dtype=int)
-    correct_shape = (n_shots, n_bits)
-    correct_counts = Counter({(0,) * n_bits: n_shots})
-    # BackendResult
-    assert np.array_equal(res.get_shots(), correct_shots)
-    assert res.get_shots().shape == correct_shape
-    assert res.get_counts() == correct_counts
-
-    # Direct
-    res = quantinuum_backend.run_circuit(c, n_shots=n_shots)
-    assert np.array_equal(res.get_shots(), correct_shots)
-    assert res.get_shots().shape == correct_shape
-    assert res.get_counts() == correct_counts
-
-
-@pytest.mark.skip("Will be moved to a set of long running integration tests")
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 @pytest.mark.parametrize(
     "authenticated_quum_backend", [{"device_name": "H2-1E"}], indirect=True
@@ -592,7 +459,6 @@ def test_retrieve_available_devices(
     assert len(backend_infos) > 0
 
 
-@pytest.mark.skip("Will be moved to a set of long running integration tests")
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 @pytest.mark.parametrize(
     "authenticated_quum_backend", [{"device_name": "H2-1E"}], indirect=True
@@ -615,7 +481,7 @@ def test_batching(
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 @pytest.mark.parametrize(
     "authenticated_quum_backend",
-    [{"device_name": name} for name in ALL_SYNTAX_CHECKER_NAMES],
+    [{"device_name": name} for name in pytest.ALL_SYNTAX_CHECKER_NAMES],
     indirect=True,
 )
 def test_submission_with_group(
@@ -715,7 +581,7 @@ def test_prefer_zzphase(
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-@pytest.mark.parametrize("device_name", ALL_DEVICE_NAMES)
+@pytest.mark.parametrize("device_name", pytest.ALL_DEVICE_NAMES)
 def test_device_state(
     device_name: str, authenticated_quum_handler: QuantinuumAPI
 ) -> None:
@@ -725,28 +591,6 @@ def test_device_state(
         ),
         str,
     )
-
-
-@pytest.mark.parametrize("device_name", ALL_DEVICE_NAMES)
-def test_defaultapi_handler(device_name: str) -> None:
-    """Test that the default API handler is used on backend construction."""
-    backend_1 = QuantinuumBackend(device_name)
-    backend_2 = QuantinuumBackend(device_name)
-
-    assert backend_1.api_handler is backend_2.api_handler
-
-
-@pytest.mark.parametrize("device_name", ALL_DEVICE_NAMES)
-def test_custom_api_handler(device_name: str) -> None:
-    """Test that custom API handlers are used when used on backend construction."""
-    handler_1 = QuantinuumAPI()
-    handler_2 = QuantinuumAPI()
-
-    backend_1 = QuantinuumBackend(device_name, api_handler=handler_1)
-    backend_2 = QuantinuumBackend(device_name, api_handler=handler_2)
-
-    assert backend_1.api_handler is not backend_2.api_handler
-    assert backend_1.api_handler._cred_store is not backend_2.api_handler._cred_store
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
@@ -795,7 +639,7 @@ def test_wasm_costs(
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 @pytest.mark.parametrize(
     "authenticated_quum_backend",
-    [{"device_name": name} for name in ALL_SYNTAX_CHECKER_NAMES],
+    [{"device_name": name} for name in pytest.ALL_SYNTAX_CHECKER_NAMES],
     indirect=True,
 )
 def test_submit_qasm(
@@ -825,7 +669,7 @@ def test_submit_qasm(
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 @pytest.mark.parametrize(
     "authenticated_quum_backend",
-    [{"device_name": name} for name in ALL_SYNTAX_CHECKER_NAMES],
+    [{"device_name": name} for name in pytest.ALL_SYNTAX_CHECKER_NAMES],
     indirect=True,
 )
 def test_options(authenticated_quum_backend: QuantinuumBackend) -> None:
@@ -843,7 +687,7 @@ def test_options(authenticated_quum_backend: QuantinuumBackend) -> None:
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 @pytest.mark.parametrize(
     "authenticated_quum_backend",
-    [{"device_name": name} for name in ALL_SYNTAX_CHECKER_NAMES],
+    [{"device_name": name} for name in pytest.ALL_SYNTAX_CHECKER_NAMES],
     indirect=True,
 )
 def test_no_opt(authenticated_quum_backend: QuantinuumBackend) -> None:
